+++ src/sessions/session-manager.ts
@@ -1,9 +1,8 @@
-
 import { GenesysCloudWebrtcSdk } from '../client';
 import BaseSessionHandler from './base-session-handler';
 import SoftphoneSessionHandler from './softphone-session-handler';
 import { LogLevels, SessionTypes, SdkErrorTypes } from '../types/enums';
-import { createAndEmitSdkError } from '../utils';
+import { createAndEmitSdkError, logPendingSession } from '../utils';
 import ScreenShareSessionHandler from './screen-share-session-handler';
 import VideoSessionHandler from './video-session-handler';
 import {
@@ -213,12 +212,12 @@ export class SessionManager {
       return;
     }
 
-    this.log('info', 'onPendingSession', sessionInfo);
+    logPendingSession(this.sdk.logger, 'onPendingSession', sessionInfo);
 
     const existingSession = this.getPendingSession(sessionInfo.sessionId);
 
     if (existingSession) {
-      this.log('info', 'duplicate session invitation, ignoring', sessionInfo);
+      logPendingSession(this.sdk.logger, 'duplicate session invitation, ignoring', sessionInfo);
       return;
     }