+++ changelog.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [PCM-1510](https://inindca.atlassian.net/browse/PCM-1510) – Be able to specify jidResource for clients (streaming client)
 * [PCM-1538](https://inindca.atlassian.net/browse/PCM-1538) – add `sdk.setAccessToken(token)` function to allow for updating the
   auth token on an already initialized sdk instance.
-* [PCM-1576](https://inindca.atlassian.net/browse/PCM-1538) - Be able to control audio volume on video and audio elements. see `sdk.updateAudioVolume`.
+* [PCM-1576](https://inindca.atlassian.net/browse/PCM-1576) - Be able to control audio volume on video and audio elements. see `sdk.updateAudioVolume`.
 
 ### Changed
 * Updated documentation.